src/requirements.txt
blinker==1.7.0
Bootstrap-Flask==2.3.3
click==8.1.7
exif==1.6.0
Flask==3.0.1
Flask-WTF==1.2.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.4
pillow==10.2.0
plum-py==0.8.7
setuptools==69.0.3
Werkzeug==3.0.1
wheel==0.42.0
WTForms==3.1.2
